refactor: moved booking validation from server to gateway
